bug fix, the columns in the rows will now update with copy
